Rename search to extract
