+++ dart_fss/fs/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
-from dart_fss.fs.search import search
+from dart_fss.fs.extract import extract
+from dart_fss.fs.fs import FinancialStatement
 
-__all__ = ['search']
+__all__ = ['extract', 'FinancialStatement']
