+++ dart_fss/tests/test_fs_search.py
@@ -1,5 +1,5 @@
 import pytest
-from dart_fss.fs.search import *
+from dart_fss.fs.extract import *
 from dart_fss.errors import NotFoundConsolidated
